requirements.txt
Flask==3.1.0
pymongo[srv]==4.15.5
dnspython==2.7.0
werkzeug==3.1.3
requests==2.32.5
gunicorn==23.0.0
itsdangerous==2.2.0
face-recognition==1.3.0
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==11.1.0
dlib==19.24.6
cryptography==44.0.0
